Move the alt text generation function into a utility file so it can be reused when needed
